Divided some components and change UI parts
