+++ CampusNavigator/CampusNavigator/CampusNavigator/Components/Common/CommonHeader.swift
@@ -1,26 +1,26 @@
-/*import SwiftUI
+import SwiftUI
 
-struct HeaderView: View {
-    var title: String
-    var rewardPoints: Int
-    
-    
+struct ResourceStatusHeader: View {
+    @Binding var rewardPoints: Int
+    @Binding var searchText: String
+    @Binding var title: String
+
     var body: some View {
         VStack(spacing: 0) {
             // Title Bar
             HStack(alignment: .center) {
-                Text(title)
+                Text("\(title)")
                     .font(.title2)
                     .fontWeight(.semibold)
                     .foregroundColor(.white)
                     .frame(maxWidth: .infinity, alignment: .leading)
                     .padding(.leading)
-                
+
                 HStack(spacing: 6) {
                     Image(systemName: "star.fill")
                         .symbolRenderingMode(.palette)
                         .foregroundStyle(.white, .green)
-                    
+
                     Text("\(rewardPoints)")
                         .fontWeight(.medium)
                         .foregroundColor(.white)
@@ -30,24 +30,39 @@ struct HeaderView: View {
                 .background(.green.opacity(0.9))
                 .clipShape(Capsule())
                 .padding(.trailing)
-                
-                Button(action: {}) {
+
+                Button {
+                    // Notification action
+                } label: {
                     Image(systemName: "bell.badge.fill")
                         .symbolRenderingMode(.palette)
                         .foregroundStyle(.red, .white)
                         .font(.system(size: 20))
                 }
-                .padding(.trailing)
+                .padding([.trailing, .vertical])
+                .padding(.leading, -10)
             }
             .frame(height: 44)
             .padding(.top, 1)
             .padding(.bottom)
-            
+
             // Search Bar
-            SearchBar()
+            HStack {
+                HStack {
+                    Image(systemName: "magnifyingglass")
+                        .foregroundColor(.gray)
+                        .padding(.leading, 8)
+
+                    TextField("Search", text: $searchText)
+                        .padding(.vertical, 8)
+                        .padding(.trailing, 8)
+                }
+                .background(Color(.systemGray6))
+                .cornerRadius(10)
+            }
+            .padding(.horizontal)
+            .padding(.bottom)
         }
         .background(Color.customDarkGreen)
     }
 }
-
-*/
