+++ CampusNavigator/CampusNavigator/CampusNavigator/Models/ResourceModel.swift
@@ -1,4 +1,13 @@
-/*import Foundation
+import Foundation
+
+
+struct ResourceItem: Identifiable {
+    let id = UUID()
+    let title: String
+    let icon: String
+    var status: ResourceStatus
+    let accuracy: Double
+}
 
 enum ResourceStatus: String, CaseIterable {
     case available = "Available"
@@ -6,11 +15,3 @@ enum ResourceStatus: String, CaseIterable {
     case moderate = "Moderate Crowded"
     case closed = "Closed"
 }
-
-struct ResourceItem: Identifiable {
-    let id = UUID() 
-    let title: String
-    let icon: String
-    var status: ResourceStatus
-    let accuracy: String
-}*/
